dashboard api and deleting funcionality made with bug fixes

+++ backend/src/controllers/sheet/sheet.contoller.js
@@ -8,7 +8,7 @@ module.exports = {
         // const { name, type, columns[0], entries[0] } = req.body;
         const userId = req.user.userId; // Extract user ID from token
         if (!userId) {
-            throw new ApiError('User ID not found in token.');
+            throw new ApiError('User not found.');
         }
         const sheetData = req.body;
 
@@ -23,7 +23,7 @@ module.exports = {
     getAllSheets: AsyncHandler(async (req, res) => {
         const userId = req.user.userId; // Extract user ID from token
         if (!userId) {
-            throw new ApiError('User ID not found in token.');
+            throw new ApiError('User not found.');
         }
         const sheets = await sheetService.getAllSheets(userId);
         if (!sheets[0]) {
@@ -36,12 +36,12 @@ module.exports = {
                 new ApiResponse(201, sheets, 'Sheet retrieved successfully.')
             );
     }),
-    updateEntry: AsyncHandler(async(req, res)=> {
+    updateEntry: AsyncHandler(async (req, res) => {
         const userId = req.user.userId;
         const { _id } = req.params;
         const { entryId, data } = req.body;
         if (!userId) {
-            throw new ApiError('User ID not found in token.');
+            throw new ApiError('User not found.');
         }
         await sheetService.update(userId, _id, entryId, data)
         res
@@ -50,18 +50,44 @@ module.exports = {
                 new ApiResponse(201, {}, 'Entry updated successfully.')
             );
     }),
-    deleteEntry: AsyncHandler(async(req, res)=> {
+    deleteEntry: AsyncHandler(async (req, res) => {
         const userId = req.user.userId;
         const { _id } = req.params;
         const { entryId } = req.body;
-        if (!userId) {
-            throw new ApiError('User ID not found in token.');
+        if (!userId || !entryId) {
+            throw new ApiError('Record not found.');
         }
         await sheetService.deleteEntry(userId, _id, entryId)
         res
             .status(201)
             .json(
                 new ApiResponse(201, {}, 'Entry deleted successfully.')
+            );
+    }),
+    deletAllEntries: AsyncHandler(async (req, res) => {
+        const userId = req.user.userId;
+        const { _id } = req.params;
+        if (!userId) {
+            throw new ApiError('User not found.');
+        }
+        await sheetService.deleteAllEntries(userId, _id)
+        res
+            .status(201)
+            .json(
+                new ApiResponse(201, {}, 'All Entries deleted successfully.')
+            );
+    }),
+    deleteSheet: AsyncHandler(async (req, res) => {
+        const userId = req.user.userId;
+        const { _id } = req.params;
+        if (!userId) {
+            throw new ApiError('User not found.');
+        }
+        await sheetService.deleteSheet(userId, _id)
+        res
+            .status(201)
+            .json(
+                new ApiResponse(201, {}, 'Sheet deleted successfully.')
             );
     })
 }
